+++ src/Queil.Ring.Configuration/Runnables/Dotnet.cs
@@ -2,12 +2,12 @@
 
 namespace Queil.Ring.Configuration.Runnables;
 
-public class AspNetCore : CsProjRunnable
+public class Dotnet : CsProjRunnable
 {
     public List<string> Urls { get; } = [];
 
-    public override string TypeId => nameof(AspNetCore).ToLowerInvariant();
+    public override string TypeId => nameof(Dotnet).ToLowerInvariant();
 
-    public override bool Equals(object? obj) => obj is AspNetCore core && UniqueId == core.UniqueId;
+    public override bool Equals(object? obj) => obj is Dotnet dotnet && UniqueId == dotnet.UniqueId;
     public override int GetHashCode() => -576574704 + UniqueId.GetHashCode();
 }