+++ src/Queil.Ring.DotNet.Cli/CsProj/LaunchSettings/iisExpress.cs
@@ -1,9 +1,0 @@
-﻿namespace Queil.Ring.DotNet.Cli.CsProj.LaunchSettings;
-
-using System;
-
-public class iisExpress
-{
-    public required Uri applicationUrl { get; init; }
-    public required int sslPort { get; init; }
-}