+++ src/Queil.Ring.DotNet.Cli/CsProj/LaunchSettings/iisSettings.cs
@@ -1,6 +1,0 @@
-﻿namespace Queil.Ring.DotNet.Cli.CsProj.LaunchSettings;
-
-public class iisSettings
-{
-    public required iisExpress iisExpress { get; init; }
-}