+++ tests/resources/apps/webapp/Program.cs
@@ -2,5 +2,6 @@ var builder = WebApplication.CreateBuilder(args);
 var app = builder.Build();
 
 app.MapGet("/", () => "OK");
+app.MapGet("/args", () => string.Join(";", args));
 
 app.Run();